@@ -11,27 +11,27 @@
     "dependencies": {
         "@mapbox/assembly": "^1.3.0",
         "@mapbox/mapbox-gl-draw": "^1.2.0",
-        "@mapbox/mapbox-gl-geocoder": "^4.6.0",
+        "@mapbox/mapbox-gl-geocoder": "^5.0.0",
         "@turf/bbox": "^6.5.0",
         "@turf/bbox-polygon": "^6.5.0",
         "@turf/buffer": "^5.1.5",
         "core-js": "^3.6.4",
         "filepond": "^4.30.3",
         "filepond-plugin-file-validate-type": "^1.2.4",
-        "floating-vue": "^1.0.0-beta.9",
+        "floating-vue": "^2.0.0-beta.8",
         "mapbox-gl": "^2.7.0",
         "mapbox-gl-draw-rectangle-mode": "^1.0.4",
         "sass": "^1.49.7",
         "sass-loader": "^10.0.0",
         "semver-sort": "^1.0.0",
-        "vue": "^2.6.11",
-        "vue-router": "^3.5.3",
-        "vue-select": "^3.16.0"
+        "vue": "^3.2.31",
+        "vue-router": "^4.0.14",
+        "vue-select": "beta"
     },
     "devDependencies": {
-        "@vue/cli-plugin-babel": "^4.2.0",
-        "@vue/cli-plugin-eslint": "^4.2.0",
-        "@vue/cli-service": "^4.2.0",
+        "@vue/cli-plugin-babel": "^5.0.3",
+        "@vue/cli-plugin-eslint": "^5.0.3",
+        "@vue/cli-service": "^5.0.3",
         "babel-eslint": "^10.0.3",
         "eslint": "^7.1.0",
         "eslint-plugin-vue": "^7.0.1",
